Adjust home page search input for mobile
