+++ apps/web/src/components/Basenames/RegistrationSearchInput/types.tsx
@@ -1,0 +1,17 @@
+import { Chain } from 'viem';
+
+export enum RegistrationSearchInputVariant {
+  Small,
+  Large,
+}
+export type RegistrationSearchInputProps = {
+  variant: RegistrationSearchInputVariant;
+  placeholder: string;
+};
+export type SuggestionEntryProps = {
+  suggestion: string;
+  buttonClasses: string;
+  handleSelectName: (name: string) => void;
+  basenameChain: Chain;
+  iconSize: number;
+};
